Refactor: extract filter and window into separate files, clean dead code
